fix(canon): keep touched files inside the source-length budget
The source-file-length gate rejected two files: helpers.rs grew past its
already over-limit base (592 -> 599) and the new template_model.rs landed
at 385 lines. Move the Vue 2 instance-member list into
helpers/vue2_members.rs and the prop-type shape probes plus their unit
tests into props/template_model/type_shape.rs.

Also refresh the two declaration-emit snapshots that still carried the
pre-guard __VizeBooleanKey definition.
